Refactoring: move class
